split cpp and h files for input classes
