+++ src/game/game.cpp
@@ -5,7 +5,7 @@
 #include "bahnhof/resources/storage.h"
 #include "bahnhof/resources/resources.h"
 #include "bahnhof/common/gamestate.h"
-#include "bahnhof/common/input.h"
+#include "bahnhof/input/input.h"
 #include "bahnhof/common/camera.h"
 #include "bahnhof/common/timing.h"
 #include "bahnhof/ui/ui.h"
