+++ src/graphics/rendering.cpp
@@ -6,7 +6,7 @@
 #include "bahnhof/graphics/graphics.h"
 #include "bahnhof/common/gamestate.h"
 #include "bahnhof/common/camera.h"
-#include "bahnhof/common/input.h"
+#include "bahnhof/input/input.h"
 #include "bahnhof/ui/ui.h"
 #include "bahnhof/track/track.h"
 #include "bahnhof/rollingstock/rollingstock.h"
