+++ src/ui/button.cpp
@@ -6,7 +6,7 @@
 #include "bahnhof/graphics/graphics.h"
 #include "bahnhof/graphics/rendering.h"
 #include "bahnhof/common/gamestate.h"
-#include "bahnhof/common/input.h"
+#include "bahnhof/input/input.h"
 #include "bahnhof/routing/routing.h"
 #include "bahnhof/rollingstock/train.h"
